requirements.txt
# minimalist requirements
-i https://pypi.org/simple
phue
Pillow
numpy
colour-science
